backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.10.6
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.28.1
openai==1.3.7
chromadb==0.5.20
sentence-transformers==3.1.1
langchain-text-splitters>=0.3.0,<1.0.0
PyMuPDF==1.24.4
python-docx==1.1.2
Pillow==11.1.0
